requirements.txt
MouseInfo==0.1.3
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
python3-xlib==0.15
PyTweening==1.0.3
